feat: px4_msgs v1.13.3
@lorenzo-bianchi @giorgiomanca97
This is our last year's version rebased onto upstream v1.13.
Their versioning system has changed significantly to support the major changes planned for v1.14 while keeping the stable v1.13 active, so pinpointing the right versions was a bit harder this time.
I'm adding this directly as code and not as a subtree/submodule, since I think that there is no reason to complicate our repo structure like theirs if we want to build a stable system.