fix: micrortps_agent: unjoinable threads
@lorenzo-bianchi @alessandrotenaglia @giorgiomanca97
For some reason, 'ros2 launch' messes with threads.
Meaning, a thread (here the receiver) that the node starts may unexplicably block when 'ros2 launch' has to forward a signal (Ctrl-C) to your application.
The result is that any attempt to join the thread results in a deadlock, and the node cannot be closed/destroyed correctly, leaving me to SIGKILL the whole thing.
I solved this by forcing threads termination in any way possible all at once. It appears to work.
I do not know why this happens, but keep this in mind while developing nodes that spawn threads, and their launch files.