feat: Flight Stack Bringup
@lorenzo-bianchi @giorgiomanca97 @alessandrotenaglia
ATTENTION EVERYONE
Consider these the newly suggested formats for launch files, for both standalone nodes and compoents.
I have to do some changes (we're not going to use rclcpp's component containers) and I'll update you when it's time, but until then, here we are.
The main problem is to be able to include these in higher-level launch files being able to configure:
- Namespaces.
- Parameter config files, considering that for components to work namespaces have to be specified in config files (see modifications to all YAMLs here). This means that we'll have to condense these config files in bigger, top-level config files that we'll pass around, but we know how to do that (yes, thanks to RoverCREA).