feat: flight_control: works in map
@alessandrotenaglia
You can pull this in and use it, should be the definitive configuration for some time (fingers crossed).
All you need to do to fly is a static tf publisher with
--frame-id map --child-frame-id drone/odom
pick values of your choice.
Note that in RViz you will see the pose in drone/odom, since that is the direct feedback from PX4's EKF2 that we need to send from FC for evaluation, debugging and testing.
Real pose information will come from the global EKF as an Odometry.
If you need an odometry feedback for your test now, subscribe to FC's ekf2_odometry topic (or something like that) and publish an identity tf from CLI (for this you just need to specify only --frame-id and --child-frame-id, no other args), so map and drone/odom coincide and EKF2's odometry makes sense.